backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
replicate==0.34.0
httpx==0.27.0
Pillow==10.4.0
python-dotenv==1.0.1
